feat(utils): extract inline useDebounce hook to shared utilities
- Extract inline useDebounce from performance-dashboard.tsx to lib/hooks/use-debounce.ts
- Eliminate blueprint.md principle 8.1 violation (DRY principle)
- Create reusable debouncing utility for 20+ components with async operations
- Perfect Service Layer compliance with zero functional changes
- Enhanced modularity following LEGO architecture principles
